Drop default scenarios and prefixes gdp_ and pop_ from names
The functions now require a scenario argument to specify the scenarios to be returned.
The SSP2EU scenario is no longer available.
The driver prefixes are now not added by default to the variable/scenario names. The argument naming is marked as deprecated an will be dropped in the next release.
Improve documentation.